SoC moved controler logic to another folder
