+++ app.js
@@ -2,7 +2,7 @@
 var express = require('express');
 var app = express();
 var path = require('path');
-var routes = require('./routes');
+var routes = require('./api/routes');
 
 //set properties of the application
 app.set('port', 3000);
